Implemented Mining
Added a Mining page "/mines" that allows the user to click a button to increase their mining level. Can only be used when the user has energy (named fatigue). The fatigue value will be reduced when mining is done, and will increase back to its original value every 30 seconds with scheduling.