+ shop automatic item listings
When the user loads the shop, it will automatically populate all items and purchase buttons that are available to buy for that location's shop to the user.

Purchase buttons populated for their respective listings, returning the name of the item clicked.

TODO: Function handling POST from buttons, removing necessary amount of money and adding the item to the user's inventory.